move event handlers to correct templates
