+++ imports/framework/Forms/DetailsForm/Dropdown/DetailsForm.Dropdown.js
@@ -2,7 +2,7 @@ import { Template } from 'meteor/templating';
 import { TAPi18n } from 'meteor/tap:i18n';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 
-import { getValue, getKey } from '/imports/framework/Forms/DetailsForm/DetailsForm.Helpers';
+import { getValue, getKey, defaultClickHandler } from '/imports/framework/Forms/DetailsForm/DetailsForm.Helpers';
 import { getEntityTranslation } from '/imports/framework/Helpers/Helpers';
 
 import './DetailsForm.Dropdown.jade';
@@ -29,4 +29,6 @@ Template.DetailsFormDropdown.onRendered(() => {});
 
 Template.DetailsFormDropdown.onDestroyed(() => {});
 
-Template.DetailsFormDropdown.events({});
+Template.DetailsFormDropdown.events({
+  'click .input': defaultClickHandler
+});
