+++ imports/framework/Forms/DetailsForm/Text/DetailsForm.Text.js
@@ -1,6 +1,6 @@
 import { Template } from 'meteor/templating';
 
-import { getValue, getKey } from '/imports/framework/Forms/DetailsForm/DetailsForm.Helpers';
+import { getValue, getKey, defaultClickHandler } from '/imports/framework/Forms/DetailsForm/DetailsForm.Helpers';
 import { getEntityTranslation } from '/imports/framework/Helpers/Helpers';
 
 import './DetailsForm.Text.jade';
@@ -17,4 +17,6 @@ Template.DetailsFormText.onRendered(() => {});
 
 Template.DetailsFormText.onDestroyed(() => {});
 
-Template.DetailsFormText.events({});
+Template.DetailsFormText.events({
+  'click .input': defaultClickHandler
+});
